Extract and fix minters tests
